+++ libs/notification/src/lib/notification.stories.ts
@@ -1,9 +1,11 @@
-import { NgModule, Component } from '@angular/core';
+import { NgModule, Component, Input, OnChanges } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { NotificationModule } from './notification.module';
 import { NotificationService } from './notification.service';
 import { NotificationWrapperComponent } from './notification-wrapper/notification-wrapper.component';
+
+import { select, text, boolean } from '@storybook/addon-knobs';
 
 export default {
   title: 'NotificationWrapperComponent',
@@ -16,15 +18,23 @@ export default {
     <button (click)="notificationService.show(data)">TestButton</button>
   `,
 })
-class StoryComponent {
+class StoryComponent implements OnChanges {
   constructor(public notificationService: NotificationService) {}
+  @Input() type?: string;
+  @Input() title?: string;
+  @Input() description?: string;
+  @Input() closeable?: boolean;
 
-  data = {
-    description: 'Test description',
-    type: 'info',
-    title: 'Test title',
-    closeable: true,
-  };
+  data: any;
+
+  ngOnChanges() {
+    this.data = {
+      description: this.description,
+      type: this.type,
+      title: this.title,
+      closeable: this.closeable,
+    };
+  }
 }
 
 @NgModule({
@@ -40,4 +50,14 @@ export const primary = () => ({
     imports: [StoryModule],
   },
   component: StoryComponent,
+  props: {
+    type: select(
+      'Type',
+      { Info: 'info', Error: 'error', Warning: 'warning', Success: 'success' },
+      'info',
+    ),
+    title: text('Text', 'Test Title'),
+    description: text('Description', 'Test Description'),
+    closeable: boolean('Closeable', true),
+  },
 });